correção e atidade televisor adicionados

+++ Atividades1608/src/entities/Area.java
@@ -2,8 +2,8 @@ package entities;
 
 public class Area {
 
-	public int comprimento;
-	public int largura; 
+	public double comprimento;
+	public double largura;
